Moved export-related operators into exporter
